Give up on fully preventing horizontal overscroll
Highly dependent upon which lines are viewed in the TUI, and needs to
change as the user scrolls. There also needs to be a "target" horizontal
scroll offset, similar to what ranges do, so that you can restore your
desired horizontal scroll offset when the TUI scrolls back to some lines
that support that offset.

It's just way more complicated than the vertical offset, and I don't
really care that much.