Use `smol`'s global executor
It's single-threaded by default, but the number of threads is
configurable with `SMOL_THREADS`.

Unfortunately, it requires futures be `Send` and `'static`. But I wasn't
able to make `smol::LocalExecutor` as ergonomic as I wanted: you have to
pass it around as an argument, you can't put it into a `static` or
whatever, because it's `!Send` and `!Sync`. So I'm going to admit defeat
for now, and return to `smol::LocalExecutor` if I really need to.

This means the benefit of `smol` over `tokio` is smaller (no pun
intended), but I still really like `smol` so I'm gonna stick with it
until I can't (if ever; I'm not writing a web server or whatever).