Use unified cursor in range
Big footgun with `Cursor`'s `Copy` impl + `{,try_}with_rope` taking
ownership and returning a new cursor (only way to change the type
parameter). The gap index changes weren't being saved to the range,
because I was modifying them on a copy, not the ones in the range!

I eventually figured it out and fixed it in this change, but the code is
very ugly and error prone, so I'm going to come up with something better
next, starting with banning `Copy`.