Drop `indigo-btree` crate
The main win I was hoping for was storing grapheme offsets and display
widths in the "measure" or "metrics" or "summary" or whatever you wanna
call it at each node. But it turns out those are impossible to keep
track of like this because they aren't monoidal.

Pretend you have one character in each node. Independently, you might
say they are each one grapheme, and each have whatever widths. But if
they were adjacent in the same node, you might consider the combination
of them one grapheme, with a display width that is not the sum of the
characters display widths individually!

I learned a good deal attempting to build this. I'm sad to let it go.
But I don't have a good reason to keep pushing on it at this point.
Especially since the `crop` crate resolves a lot of the nit picks I have
with the `ropey` crate, so I could switch to that. And I never would've
approached the performance of either of those crates. I'm better off
contributing to them or forking them if I have unmet needs.