Use `ExitCode` type
I think I kept it a `u8` because I wanted to know what the specific code
was, since `ExitCode` hides it from you. But the docs make a good point
that the code is platform-specific. And I don't currently care about the
code. So I'll switch to `ExitCode`. If I start caring about the exit
reason, I can make a custom type to represent exit reasons for Indigo.