Delete lending iterator code
I actually prefer the ergonomics of the `for_each` methods. They work
with the pipeline style.

The lending iterators required creating a separate binding to store
iterator state. The `while` loop syntax wasn't too bad though.

Ultimately it's cool that I can do this but I don't find myself wanting
this now that I have it. So it's gotta go.